> Commit classe padre Stats

+++ MioEsame/src/main/java/com/exam/Interface/AbstactStatistic.java
@@ -14,16 +14,14 @@ import com.exam.model.Tweet;
  */
 public abstract class AbstactStatistic {
 	
-	//private Map<String, Object> map;
+	
 	
 	public AbstactStatistic() {
 		
 	}
-	
-   // public abstract void Stats(Map<String, Object> map, boolean numbers);
 		
 	
-	public abstract int calcola(ArrayList<Tweet> vett, String field);
+	public abstract double calcola(ArrayList<Tweet> vett, String field);
